chore: remove deprecated tools
