+++ src/rai/rai/tools/time.py
@@ -14,8 +14,11 @@
 #
 
 import time
+from typing import Type
 
+from langchain.pydantic_v1 import BaseModel, Field
 from langchain.tools import tool
+from langchain_core.tools import BaseTool
 
 
 @tool
@@ -25,3 +28,27 @@ def sleep_max_5s(n: int):
         n = 5
 
     time.sleep(n)
+
+
+class WaitForSecondsToolInput(BaseModel):
+    """Input for the WaitForSecondsTool tool."""
+
+    seconds: int = Field(..., description="The number of seconds to wait")
+
+
+class WaitForSecondsTool(BaseTool):
+    """Wait for a specified number of seconds"""
+
+    name: str = "WaitForSecondsTool"
+    description: str = (
+        "A tool for waiting. "
+        "Useful for pausing execution for a specified number of seconds. "
+        "Input should be the number of seconds to wait."
+    )
+
+    args_schema: Type[WaitForSecondsToolInput] = WaitForSecondsToolInput
+
+    def _run(self, seconds: int):
+        """Waits for the specified number of seconds."""
+        time.sleep(seconds)
+        return f"Waited for {seconds} seconds."
